refactor: CompanyQueryCondition 패키지 변경
