+++ company-service/src/main/java/com/firstlogistics/companyservice/domain/repository/CompanyRepository.java
@@ -1,6 +1,6 @@
 package com.firstlogistics.companyservice.domain.repository;
 
-import com.firstlogistics.companyservice.application.dto.query.CompanyQueryCondition;
+import com.firstlogistics.companyservice.domain.repository.dto.CompanyQueryCondition;
 import com.firstlogistics.companyservice.domain.entity.Company;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
